Move all decoding code under the respective chain modules
