--- rotkehlchen/chain/ethereum/decoding/liquity/__init__.py
@@ -1,1 +1,0 @@
-from .decoder import LiquityDecoder  # noqa: F401
