--- rotkehlchen/chain/ethereum/decoding/makerdao/__init__.py
@@ -1,1 +1,0 @@
-from .decoder import MakerdaoDecoder  # noqa: F401
