--- rotkehlchen/chain/ethereum/decoding/pickle/__init__.py
@@ -1,1 +1,0 @@
-from .decoder import PickleDecoder  # noqa: F401
