+++ rotkehlchen/chain/ethereum/modules/airdrops/decoder.py
