+++ rotkehlchen/chain/ethereum/modules/compound/__init__.py
@@ -1,1 +1,2 @@
+from .compound import Compound  # noqa: F401
 from .decoder import CompoundDecoder  # noqa: F401
