+++ rotkehlchen/chain/ethereum/modules/curve/decoder.py
