+++ rotkehlchen/chain/ethereum/modules/dxdaomesa/data/contracts.json
