+++ rotkehlchen/chain/ethereum/modules/dxdaomesa/decoder.py
