+++ rotkehlchen/chain/ethereum/modules/ens/decoder.py
