+++ rotkehlchen/chain/ethereum/modules/gitcoin/decoder.py
