+++ rotkehlchen/chain/ethereum/modules/kyber/decoder.py
