+++ rotkehlchen/chain/ethereum/modules/liquity/__init__.py
@@ -1,0 +1,1 @@
+from .decoder import LiquityDecoder  # noqa: F401
