+++ rotkehlchen/chain/ethereum/modules/makerdao/__init__.py
@@ -1,4 +1,5 @@
 __all__ = ['MakerdaoDsr', 'MakerdaoVaults']
 
-from rotkehlchen.chain.ethereum.modules.makerdao.dsr import MakerdaoDsr
-from rotkehlchen.chain.ethereum.modules.makerdao.vaults import MakerdaoVaults
+from .decoder import MakerdaoDecoder  # noqa: F401
+from .dsr import MakerdaoDsr
+from .vaults import MakerdaoVaults
