+++ rotkehlchen/chain/ethereum/modules/oneinch/v1/decoder.py
