+++ rotkehlchen/chain/ethereum/modules/oneinch/v2/decoder.py
