+++ rotkehlchen/chain/ethereum/modules/pickle/__init__.py
@@ -1,1 +1,2 @@
+from .decoder import PickleDecoder  # noqa: F401
 from .pickle import PickleFinance  # noqa: F401
