+++ rotkehlchen/chain/ethereum/modules/pickle/decoder.py
