+++ rotkehlchen/chain/ethereum/modules/uniswap/v1/decoder.py
