+++ rotkehlchen/chain/ethereum/modules/votium/decoder.py
