+++ rotkehlchen/chain/ethereum/modules/zksync/decoder.py
